requirements.txt
Flask>=2.0.0
Pillow>=9.0.0
pyinstaller>=5.0
